+++ RemoteControl/CHANGELOG.md
@@ -13,6 +13,10 @@ All notable changes to this RDK Service will be documented in this file.
 * For more details, refer to [versioning](https://github.com/rdkcentral/rdkservices#versioning) section under Main README.
 * The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.3] - 2023-01-26
+### Changed
+- RDK-44991: Upgrade Flex-2.0 devices to use Thunder R4.4.1
+
 ## [1.4.2] - 2023-12-13
 ### Changed
 - Remove remoteControlTestClient from build